Move error extraction code to more suitable place, and make tests reusable.
