+++ eliot/tests/test_action.py
@@ -1252,107 +1252,113 @@ class WrittenActionTests(testtools.TestCase):
             WrittenAction.from_messages, start_message, messages)
 
 
-class FailedActionFieldExtraction(TestCase):
+def make_error_extraction_tests(get_messages):
     """
-    Tests for extracting fields from exceptions in failed actions.
+    Create a test case class for testing extraction of fields from exceptions.
+
+    @param get_messages: Callable that takes an exception instance, returns
+        all message dictionaries generated by logging it.
+
+    @return: ``TestCase`` subclass.
     """
-    def get_failed_action_message(self, exception):
+    class ErrorFieldExtraction(TestCase):
         """
-        Fail an action using the given exception.
-
-        :return: Logged dictionary from the exception failing an action.
+        Tests for extracting fields from exceptions in failed actions.
         """
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", TaskLevel(level=[1]), "sys:me")
-        try:
-            with action:
-                raise exception
-        except exception.__class__:
-            pass
-        return logger.messages[-1]
+        def test_matching_class(self):
+            """
+            If an exception fails an action and the exact type has registered
+            extractor, extract errors using it.
+            """
+            class MyException(Exception):
+                pass
+            extract_fields_for_failures(MyException,
+                                        lambda e: {"key": e.args[0]})
+            [message] = get_messages(MyException("a value"))
+            assertContainsFields(self, message, {"key": "a value"})
 
-    def test_matching_class(self):
-        """
-        If an exception fails an action and the exact type has registered
-        extractor, extract errors using it.
-        """
-        class MyException(Exception):
-            pass
-        extract_fields_for_failures(MyException,
-                                    lambda e: {"key": e.args[0]})
-        message = self.get_failed_action_message(MyException("a value"))
-        assertContainsFields(self, message, {"key": "a value"})
+        def test_subclass_falls_back_to_parent(self):
+            """
+            If an exception fails an action and the exact type has not been
+            registered but the error is a subclass of a registered class,
+            extract errors using it.
+            """
+            class MyException(Exception):
+                pass
 
-    def test_subclass_falls_back_to_parent(self):
-        """
-        If an exception fails an action and the exact type has not been
-        registered but the error is a subclass of a registered class,
-        extract errors using it.
-        """
-        class MyException(Exception):
-            pass
+            class SubException(MyException):
+                pass
 
-        class SubException(MyException):
-            pass
+            extract_fields_for_failures(MyException,
+                                        lambda e: {"key": e.args[0]})
+            [message] = get_messages(SubException("the value"))
+            assertContainsFields(self, message, {"key": "the value"})
 
-        extract_fields_for_failures(MyException,
-                                    lambda e: {"key": e.args[0]})
-        message = self.get_failed_action_message(SubException("the value"))
-        assertContainsFields(self, message, {"key": "the value"})
+        def test_subclass_matches_first(self):
+            """
+            If both a superclass and base class have registered extractors, the
+            more specific one is used.
+            """
+            class MyException(Exception):
+                pass
 
-    def test_subclass_matches_first(self):
-        """
-        If both a superclass and base class have registered extractors, the
-        more specific one is used.
-        """
-        class MyException(Exception):
-            pass
+            class SubException(MyException):
+                pass
 
-        class SubException(MyException):
-            pass
+            class SubSubException(SubException):
+                pass
 
-        class SubSubException(SubException):
-            pass
+            extract_fields_for_failures(MyException,
+                                        lambda e: {"parent": e.args[0]})
+            extract_fields_for_failures(SubException,
+                                        lambda e: {"child": e.args[0]})
+            [message] = get_messages(SubSubException("the value"))
+            assertContainsFields(self, message, {"child": "the value"})
 
-        extract_fields_for_failures(MyException,
-                                    lambda e: {"parent": e.args[0]})
-        extract_fields_for_failures(MyException,
-                                    lambda e: {"child": e.args[0]})
-        message = self.get_failed_action_message(SubSubException("the value"))
-        assertContainsFields(self, message, {"child": "the value"})
+        def test_error_in_extracter(self):
+            """
+            If an error occurs in extraction, log the message as usual just
+            without the extra fields, and an additional traceback.
+            """
+            class MyException(Exception):
+                pass
 
-    def test_error_in_extracter(self):
-        """
-        If an error occurs in extraction, log the message as usual just
-        without the extra fields, and an additional traceback.
-        """
-        class MyException(Exception):
-            pass
+            def extract(e):
+                return e.nosuchattribute
+            extract_fields_for_failures(MyException, extract)
 
-        def extract(e):
-            return e.nosuchattribute
-        extract_fields_for_failures(MyException, extract)
+            messages = get_failed_action_messages(MyException())
+            assertContainsFields(self, messages[1],
+                                 {"action_type": "sys:me",
+                                  "action_status": "failed"})
+            assertContainsFields(self, messages[0],
+                                 {"message_type": "eliot:traceback"})
+            self.assertIn("nosuchattribute", str(messages[0]["reason"]))
 
-        logger = MemoryLogger()
-        action = Action(logger, "uuid", TaskLevel(level=[1]), "sys:me")
-        try:
-            with action:
-                raise MyException()
-        except MyException:
-            pass
+        def test_environmenterror(self):
+            """
+            ``EnvironmentError`` has a registered extractor that extracts the
+            errno.
+            """
+            [message] = get_messages(EnvironmentError(12, "oh noes"))
+            assertContainsFields(self, message, {"errno": 12})
 
-        assertContainsFields(self, logger.messages[1],
-                             {"action_type": "sys:me",
-                              "action_status": "failed"})
-        assertContainsFields(self, logger.messages[0],
-                             {"message_type": "eliot:traceback"})
-        self.assertIn("nosuchattribute", str(logger.messages[0]["reason"]))
+    return ErrorFieldExtraction
 
-    def test_environmenterror(self):
-        """
-        ``EnvironmentError`` has a registered extractor that extracts the
-        errno.
-        """
-        message = self.get_failed_action_message(
-            EnvironmentError(12, "oh noes"))
-        assertContainsFields(self, message, {"errno": 12})
+
+def get_failed_action_messages(exception):
+    """
+    Fail an action using the given exception.
+
+    :return: Logged dictionaries from the exception failing an action.
+    """
+    logger = MemoryLogger()
+    action = Action(logger, "uuid", TaskLevel(level=[1]), "sys:me")
+    try:
+        with action:
+            raise exception
+    except exception.__class__:
+        pass
+    return logger.messages
+FailedActionExtractionTests = make_error_extraction_tests(
+    get_failed_action_messages)
